+++ docs/cost_sensitive_multiclass.rst
