+++ libact/query_strategies/multiclass/__init__.py
@@ -1,0 +1,1 @@
+from .active_learning_with_cost_embedding import ActiveLearningWithCostEmbedding
